Multiple updates
- Updated UI: added the Line text edit to set numeric values
- Completes the main part of the SW.
After beginning testing, I understood that I was wrong with the manner of using
QSemaphor to organize atomic protection. We can't stop on one slot (button click processing)
and wait for another (reply from the serial communication) on the same thread.
In this step, I have to back to the callback manner of passing the results of processing
communication to the UI thread.